playlist into seperate component
