+++ src/components/playlist/playlist.css
@@ -1,0 +1,49 @@
+
+.Playlist {
+  width: 85%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  overflow: scroll;
+  margin: 0;
+  padding: 0;
+
+  li {
+    list-style: none;
+    display: grid;
+    grid-template-columns: 25% 1fr 8%;
+    width: 100%;
+    background: var(--off-white);
+    margin-bottom: 1em;
+
+    h2 {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+    }
+
+    h3 {
+      display: flex;
+      align-items: center;
+      text-align: center;
+      overflow: hidden;
+      padding: 1em;
+    }
+
+    div {
+      background: red;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      font-family: var(--paragraph-font);
+      font-weight: bolder;
+      color: var(--dark-blue);
+    }
+  }
+}
+
+@media screen and (min-width: 600px) {
+  .Playlist {
+    width: 60%;
+  }
+}
